// src/app/software/model/trn-sold-unit-requirement-activity.model.ts
export class TrnSoldUnitRequirementActivityModel {
    Id: number = 0;
    SoldUnitRequirementId: number = 0;
    ActivityDate: string = "";
    Activity: string = "";
    Remarks: string = "";
    UserId: number = 0;
    User: string = "";
    ChecklistRequirement: string = "";
    SoldUnitNumber: string = "";
    Project: string = "";
    UnitCode: string = "";
    Customer: string = "";
}